+++ app/src/main/java/os/com/ui/contest/activity/ContestDetailActivity.kt
@@ -84,7 +84,7 @@ class ContestDetailActivity : BaseActivity(), View.OnClickListener, OnClickRecyc
                     contest_id = this.contest_id
                 else contest_id = contest!!.contest_id
 
-                if (!data!!.total_winners.isEmpty() && data!!.total_winners.toInt() > 0)
+                if (!data!!.total_winners.isEmpty() && data!!.total_winners.toLong() > 0)
                     callWinningBreakupApi(
                         contest_id,
                         data!!.breakup_detail!!,
@@ -276,7 +276,7 @@ class ContestDetailActivity : BaseActivity(), View.OnClickListener, OnClickRecyc
 
         if (!data.total_teams.isEmpty() && !data.teams_joined.isEmpty()) {
             val strtValue =
-                data.total_teams.toLong() - data.teams_joined.toInt()
+                data.total_teams.toLong() - data.teams_joined.toLong()
             txt_StartValue.text = getString(R.string.only) + " " + strtValue.toString() + " " +
                     getString(R.string.spots_left)
             crs_Progress.setMinValue(0f)
@@ -361,7 +361,7 @@ class ContestDetailActivity : BaseActivity(), View.OnClickListener, OnClickRecyc
             if (data.is_joined) {
                 cl_join.visibility = GONE
                 cl_viewJoined.visibility = VISIBLE
-                val total_teams = data.total_teams.toLong() - data.teams_joined.toInt()
+                val total_teams = data.total_teams.toLong() - data.teams_joined.toLong()
                 if (total_teams > 0) {
                     ll_bottom.visibility = VISIBLE
                     btn_InviteFriends.visibility = VISIBLE
@@ -385,7 +385,7 @@ class ContestDetailActivity : BaseActivity(), View.OnClickListener, OnClickRecyc
                 cl_join.visibility = GONE
                 cl_viewJoined.visibility = VISIBLE
                 txt_Join.text = getString(R.string.join_plus)
-                val total_teams = data.total_teams.toLong() - data.teams_joined.toInt()
+                val total_teams = data.total_teams.toLong() - data.teams_joined.toLong()
                 if (total_teams > 0) {
                     ll_bottom.visibility = VISIBLE
                     btn_InviteFriends.visibility = VISIBLE
